+++ Tum4ik.JustClipboardManager/Properties/SettingsPasteWindow.Designer.cs
@@ -34,5 +34,17 @@ namespace Tum4ik.JustClipboardManager.Properties {
                 this["SnappingType"] = value;
             }
         }
+        
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("0")]
+        public int DisplayCorner {
+            get {
+                return ((int)(this["DisplayCorner"]));
+            }
+            set {
+                this["DisplayCorner"] = value;
+            }
+        }
     }
 }