+++ ReleaseNotes.md
@@ -2,6 +2,7 @@
 * Plugins support.
 * Keep user settings from the previous version of the application after upgrade.
 * Automatic restart on application crash.
+* Paste window settings.
 
 ### Fixes
 * Fixed problem when the application breaks the clipboard viewer chain on exit.